winfo: account noise for multiple radio from same wiphy
With the introduction of MLO, a wiphy might have multiple radio serving
it. The current code that parse the noise level use the IN_USE flag to
identify the affected radio but this changed with MLO as now multiple
radio can have IN_USE flag set.

This confuse the parsing as only the first radio is parsed with all the
other skipped, resulting in all the radio for the wiphy wrongly reporting
the noise of the first.

Rework the noise parsing logic to check if the frequency match the
current interface in addition to the IN_USE flag.

Suggested-by: Valeriy Kosikhin <vkosikhin@gmail.com>
Tested-by: Aleksander Jan Bajkowski <olek2@wp.pl> # tested on Quantum W1700k
Co-developed-by: Valeriy Kosikhin <vkosikhin@gmail.com>
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>
(cherry picked from commit 904f114548bac1e21f49f0bf1e16ebd8ac907885)